+++ app/config/cms.php
@@ -35,6 +35,20 @@ return array(
 
     /*
     |--------------------------------------------------------------------------
+    | Prevents application updates
+    |--------------------------------------------------------------------------
+    |
+    | If using composer or git to download updates to the core files, set this
+    | value to 'true' to prevent the update gateway from trying to download
+    | these files again as part of the application update process. Plugins
+    | and themes will still be downloaded.
+    |
+    */
+
+    'disableCoreUpdates' => false,
+
+    /*
+    |--------------------------------------------------------------------------
     | Back-end URI prefix
     |--------------------------------------------------------------------------
     |