+++ modules/system/controllers/Updates.php
@@ -4,6 +4,7 @@ use Str;
 use Lang;
 use File;
 use Flash;
+use Config;
 use Backend;
 use Redirect;
 use BackendMenu;
@@ -31,6 +32,11 @@ class Updates extends Controller
 
     public $listConfig = ['list' => 'config_list.yaml', 'manage' => 'config_manage_list.yaml'];
 
+    /**
+     * @var boolean If set to true, core updates will not be downloaded or extracted.
+     */
+    protected $disableCoreUpdates = false;
+
     public function __construct()
     {
         parent::__construct();
@@ -38,6 +44,8 @@ class Updates extends Controller
         $this->addCss('/modules/system/assets/css/updates.css', 'core');
 
         BackendMenu::setContext('October.System', 'system', 'updates');
+
+        $this->disableCoreUpdates = Config::get('cms.disableCoreUpdates', false);
     }
 
     /**
@@ -98,10 +106,12 @@ class Updates extends Controller
 
         switch ($stepCode) {
             case 'downloadCore':
+                if ($this->disableCoreUpdates) return;
                 $manager->downloadCore(post('hash'));
                 break;
 
             case 'extractCore':
+                if ($this->disableCoreUpdates) return;
                 $manager->extractCore(post('hash'), post('build'));
                 break;
 
@@ -109,8 +119,16 @@ class Updates extends Controller
                 $manager->downloadPlugin(post('name'), post('hash'));
                 break;
 
+            case 'downloadTheme':
+                $manager->downloadTheme(post('name'), post('hash'));
+                break;
+
             case 'extractPlugin':
                 $manager->extractPlugin(post('name'), post('hash'));
+                break;
+
+            case 'extractTheme':
+                $manager->extractTheme(post('name'), post('hash'));
                 break;
 
             case 'completeUpdate':
@@ -148,7 +166,8 @@ class Updates extends Controller
 
             $this->vars['core'] = array_get($result, 'core', false);
             $this->vars['hasUpdates'] = array_get($result, 'update', false);
-            $this->vars['updateList'] = array_get($result, 'plugins', []);
+            $this->vars['pluginList'] = array_get($result, 'plugins', []);
+            $this->vars['themeList'] = array_get($result, 'themes', []);
         }
         catch (Exception $ex) {
             $this->handleError($ex);
@@ -176,10 +195,16 @@ class Updates extends Controller
                 $plugins[$code] = array_get($plugin, 'hash', null);
             }
 
+            $themes = [];
+            $themeList = array_get($result, 'themes', []);
+            foreach ($themeList as $code => $theme) {
+                $themes[$code] = array_get($theme, 'hash', null);
+            }
+
             /*
              * Update steps
              */
-            $updateSteps = $this->buildUpdateSteps($core, $plugins);
+            $updateSteps = $this->buildUpdateSteps($core, $plugins, $themes);
 
             /*
              * Finish up
@@ -204,18 +229,20 @@ class Updates extends Controller
     public function onApplyUpdates()
     {
         try {
-            $plugins = post('plugins', []);
-            if (!is_array($plugins))
-                $plugins = [];
-
             $coreHash = post('hash');
             $coreBuild = post('build');
             $core = [$coreHash, $coreBuild];
 
+            $plugins = post('plugins', []);
+            if (!is_array($plugins)) $plugins = [];
+
+            $themes = post('themes', []);
+            if (!is_array($themes)) $themes = [];
+
             /*
              * Update steps
              */
-            $updateSteps = $this->buildUpdateSteps($core, $plugins);
+            $updateSteps = $this->buildUpdateSteps($core, $plugins, $themes);
 
             /*
              * Finish up
@@ -234,13 +261,16 @@ class Updates extends Controller
         return $this->makePartial('execute');
     }
 
-    private function buildUpdateSteps($core, $plugins)
+    private function buildUpdateSteps($core, $plugins, $themes)
     {
+        if (!is_array($core))
+            $core = [null, null];
+
         if (!is_array($plugins))
             $plugins = [];
 
-        if (!is_array($core))
-            $core = [null, null];
+        if (!is_array($themes))
+            $themes = [];
 
         $updateSteps = [];
         list($coreHash, $coreBuild) = $core;
@@ -265,6 +295,15 @@ class Updates extends Controller
             ];
         }
 
+        foreach ($themes as $name => $hash) {
+            $updateSteps[] = [
+                'code'  => 'downloadTheme',
+                'label' => Lang::get('system::lang.updates.theme_downloading', compact('name')),
+                'name'  => $name,
+                'hash'  => $hash
+            ];
+        }
+
         /*
          * Extract
          */
@@ -281,6 +320,15 @@ class Updates extends Controller
             $updateSteps[] = [
                 'code' => 'extractPlugin',
                 'label' => Lang::get('system::lang.updates.plugin_extracting', compact('name')),
+                'name' => $name,
+                'hash' => $hash
+            ];
+        }
+
+        foreach ($themes as $name => $hash) {
+            $updateSteps[] = [
+                'code' => 'extractTheme',
+                'label' => Lang::get('system::lang.updates.theme_extracting', compact('name')),
                 'name' => $name,
                 'hash' => $hash
             ];